used ReactContext on Link component
