+++ web/src/components/authors/index.tsx
@@ -1,4 +1,5 @@
 import { GithubUser } from "@dzcode.io/api/dist/app/types/legacy";
+import { Link } from "@dzcode.io/ui/dist/link";
 import { TryAgain } from "@dzcode.io/ui/dist/try-again";
 import { LOADABLE } from "@dzcode.io/utils/dist/loadable";
 import Avatar from "@material-ui/core/Avatar";
@@ -7,7 +8,6 @@ import Tooltip from "@material-ui/core/Tooltip";
 import Typography from "@material-ui/core/Typography";
 import Skeleton from "@material-ui/lab/Skeleton";
 import { FC } from "react";
-import { LinkV2 } from "src/components/link-v2";
 import { fetchCurrentArticleAuthors } from "src/redux/actions/articles-page";
 
 const useStyles = makeStyles((theme) => ({
@@ -44,11 +44,11 @@ export const Authors: FC<AuthorsProps> = ({ githubAuthors }) => {
           <div className={classes.avatarsContainer}>
             {githubAuthors
               ? githubAuthors.map((author, index) => (
-                  <LinkV2 key={`author-${index}`} className={classes.avatar} href={author.html_url}>
+                  <Link key={`author-${index}`} className={classes.avatar} href={author.html_url}>
                     <Tooltip title={author.login} aria-label={author.login}>
                       <Avatar src={author.avatar_url} />
                     </Tooltip>
-                  </LinkV2>
+                  </Link>
                 ))
               : [1, 2, 3].map((item, index) => (
                   <Skeleton
