+++ web/src/components/contributors/index.tsx
@@ -1,4 +1,5 @@
 import { GithubUser } from "@dzcode.io/api/dist/app/types/legacy";
+import { Link } from "@dzcode.io/ui/dist/link";
 import { TryAgain } from "@dzcode.io/ui/dist/try-again";
 import { LOADABLE } from "@dzcode.io/utils/dist/loadable";
 import Avatar from "@material-ui/core/Avatar";
@@ -7,7 +8,6 @@ import Tooltip from "@material-ui/core/Tooltip";
 import Typography from "@material-ui/core/Typography";
 import Skeleton from "@material-ui/lab/Skeleton";
 import { FC } from "react";
-import { LinkV2 } from "src/components/link-v2";
 import { fetchCurrentArticleContributors } from "src/redux/actions/articles-page";
 
 const useStyles = makeStyles((theme) => ({
@@ -44,7 +44,7 @@ export const Contributors: FC<ContributorsProps> = ({ contributors }) => {
           <div className={classes.avatarsContainer}>
             {contributors
               ? contributors.map((contributor, index) => (
-                  <LinkV2
+                  <Link
                     key={`contributor-${index}`}
                     className={classes.avatar}
                     href={contributor.html_url}
@@ -52,7 +52,7 @@ export const Contributors: FC<ContributorsProps> = ({ contributors }) => {
                     <Tooltip title={contributor.login} aria-label={contributor.login}>
                       <Avatar src={contributor.avatar_url} />
                     </Tooltip>
-                  </LinkV2>
+                  </Link>
                 ))
               : [1, 2, 3].map((item, index) => (
                   <Skeleton
