+++ web/src/components/markdown/index.tsx
@@ -1,3 +1,4 @@
+import { Link } from "@dzcode.io/ui/dist/link";
 import useTheme from "@material-ui/core/styles/useTheme";
 import Typography from "@material-ui/core/Typography";
 import ReactMarkdown, { MarkdownToJSX } from "markdown-to-jsx";
@@ -5,7 +6,6 @@ import { FC } from "react";
 import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
 import prism from "react-syntax-highlighter/dist/cjs/styles/prism/prism";
 import tomorrow from "react-syntax-highlighter/dist/cjs/styles/prism/tomorrow";
-import { LinkV2 } from "src/components/link-v2";
 import { useSliceSelector } from "src/redux/selectors";
 
 export const Markdown: FC<ReactMarkdown> = ({ children, t, ...markdownProps }) => {
@@ -37,7 +37,7 @@ export const Markdown: FC<ReactMarkdown> = ({ children, t, ...markdownProps }) =
             props: { gutterBottom: true, variant: "caption", paragraph: true },
           },
           p: { component: Typography, props: { paragraph: true } },
-          a: { component: LinkV2 },
+          a: { component: Link },
           img: { props: { style: { maxWidth: "100%" } } },
           code: {
             props: {
