+++ web/src/pages/not-found/index.tsx
@@ -1,4 +1,5 @@
 import { ErrorBoundary } from "@dzcode.io/ui/dist/error-boundary";
+import { Link } from "@dzcode.io/ui/dist/link";
 import Button from "@material-ui/core/Button";
 import makeStyles from "@material-ui/core/styles/makeStyles";
 import Typography from "@material-ui/core/Typography";
@@ -7,7 +8,6 @@ import ArrowForwardIcon from "@material-ui/icons/ArrowForward";
 import { FC } from "react";
 import { Helmet } from "react-helmet";
 import svg from "src/assets/svg/404.svg";
-import { LinkV2 } from "src/components/link-v2";
 import { Markdown } from "src/components/markdown";
 import { T, t } from "src/components/t";
 import { useSliceSelector } from "src/redux/selectors";
@@ -42,14 +42,14 @@ const NotFound: FC = () => {
         <Typography className={classes.text}>
           <Markdown t={t("notfound-subtitle")} />
         </Typography>
-        <LinkV2 href="/">
+        <Link href="/">
           <Button
             startIcon={language.code === "ar" ? <ArrowForwardIcon /> : <ArrowBackIcon />}
             size="large"
           >
             <T notfound-back-home />
           </Button>
-        </LinkV2>
+        </Link>
       </div>
     </ErrorBoundary>
   );
